Implement LayoutOptionsResolver
