+++ Sources/StreamChatUI/ChatMessageList/RETHINK/MessageThreadVC.swift
@@ -36,6 +36,8 @@ class MessageThreadVC<ExtraData: ExtraDataTypes>: _ViewController, UICollectionV
         .messageComposerViewController
         .init()
     
+    public private(set) lazy var layoutOptionsResolver = uiConfig.messageList.layoutOptionsResolver
+    
     private var messageComposerBottomConstraint: NSLayoutConstraint?
     
     private lazy var keyboardObserver = ChatMessageListKeyboardObserver(
@@ -135,76 +137,6 @@ class MessageThreadVC<ExtraData: ExtraDataTypes>: _ViewController, UICollectionV
         return delay > minTimeIntervalBetweenMessagesInGroup
     }
     
-    // It's not using isPartOfThread
-    func cellLayoutOptionsForMessage(at indexPath: IndexPath) -> ChatMessageLayoutOptions {
-        let message = chatMessage(for: indexPath)
-        let isLastInGroup = isMessageLastInGroup(at: indexPath)
-
-        var options: ChatMessageLayoutOptions = []
-
-        if message.isSentByCurrentUser {
-            options.insert(.flipped)
-        }
-        if !isLastInGroup {
-            options.insert(.continuousBubble)
-        }
-        if !isLastInGroup && !message.isSentByCurrentUser {
-            options.insert(.avatarSizePadding)
-        }
-        if isLastInGroup {
-            options.insert(.metadata)
-        }
-        if !message.textContent.isEmpty {
-            options.insert(.text)
-        }
-
-        guard message.deletedAt == nil else {
-            return options
-        }
-
-        if isLastInGroup && !message.isSentByCurrentUser {
-            options.insert(.avatar)
-        }
-        if message.quotedMessageId != nil {
-            options.insert(.quotedMessage)
-        }
-        if !message.reactionScores.isEmpty {
-            options.insert(.reactions)
-        }
-        if message.lastActionFailed {
-            options.insert(.error)
-        }
-
-        let attachmentOptions: ChatMessageLayoutOptions = message.attachments.reduce([]) { options, attachment in
-            if (attachment as? ChatMessageDefaultAttachment)?.actions.isEmpty == false {
-                return options.union(.actions)
-            }
-
-            switch attachment.type {
-            case .image:
-                return options.union(.photoPreview)
-            case .giphy:
-                return options.union(.giphy)
-            case .file:
-                return options.union(.filePreview)
-            case .link:
-                return options.union(.linkPreview)
-            default:
-                return options
-            }
-        }
-
-        if attachmentOptions.contains(.actions) {
-            options.insert(.actions)
-        } else if attachmentOptions.intersection([.photoPreview, .giphy, .filePreview]).isEmpty == false {
-            options.formUnion(attachmentOptions.subtracting(.linkPreview))
-        } else if attachmentOptions.contains(.linkPreview) {
-            options.insert(.linkPreview)
-        }
-
-        return options
-    }
-    
     func cellReuseIdentifier(for message: _ChatMessage<ExtraData>) -> String {
         MessageCell<ExtraData>.reuseId
     }
@@ -225,7 +157,9 @@ class MessageThreadVC<ExtraData: ExtraDataTypes>: _ViewController, UICollectionV
         let message = chatMessage(for: indexPath)
         
         let reuseId = cellReuseIdentifier(for: message)
-        let layoutOptions = cellLayoutOptionsForMessage(at: indexPath)
+        let isLastInGroup = isMessageLastInGroup(at: indexPath)
+        var layoutOptions = layoutOptionsResolver(message, isLastInGroup)
+        layoutOptions.remove(.threadInfo)
         
         let cell: MessageCell<ExtraData> = self.collectionView.dequeueReusableCell(
             withReuseIdentifier: reuseId,
